ПР14_ревью1

+++ controllers/cards.js
@@ -1,46 +1,45 @@
 const http2 = require('http2');
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-err');
+const BadRequestError = require('../errors/bad-request-err');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .populate(['owner', 'likes'])
     .then((cards) => res.status(http2.constants.HTTP_STATUS_OK).send({ data: cards }))
-    .catch(() => res.status(http2.constants.NGHTTP2_INTERNAL_ERROR).send({ message: 'Произошла ошибка при запросе всех карточек' }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
   Card.create({ name, link, owner })
     .then((card) => res.status(http2.constants.HTTP_STATUS_OK).send({ data: card }))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        res.status(http2.constants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Переданы некорректные данные при создании карточки' });
+    .catch(next);
+};
+
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .then((delCard) => {
+      const cardOwnerId = delCard.owner._id.toString();
+      const userId = req.user._id.toString();
+      if (cardOwnerId === userId) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((card) => {
+            if (!card) {
+              throw new NotFoundError('Карточка c указанным id не найдена');
+            }
+            return res.status(http2.constants.HTTP_STATUS_OK).send({ data: card });
+          })
+          .catch(next);
       } else {
-        res.status(http2.constants.NGHTTP2_INTERNAL_ERROR).send({ message: 'Ошибка по умолчанию' });
+        throw new BadRequestError('Нельзя удалить чужую карточку');
       }
     });
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .then((card) => {
-      if (!card) {
-        return res.status(http2.constants.HTTP_STATUS_NOT_FOUND).send({ message: 'Карточка c указанным id не найдена' });
-      }
-      return res.status(http2.constants.HTTP_STATUS_OK).send({ data: card });
-    })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(http2.constants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Переданы некорректные данные карточки' });
-      } else {
-        res.status(http2.constants.NGHTTP2_INTERNAL_ERROR).send({ message: 'Ошибка по умолчанию' });
-      }
-    });
-};
-
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -48,19 +47,14 @@ module.exports.likeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        return res.status(http2.constants.HTTP_STATUS_NOT_FOUND).send({ message: 'Передан несуществующий _id карточки.' });
+        throw new NotFoundError('Передан несуществующий _id карточки.');
       }
       return res.status(http2.constants.HTTP_STATUS_OK).send({ data: card });
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(http2.constants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Переданы некорректные данные для постановки/снятия лайка.' });
-      }
-      return res.status(http2.constants.NGHTTP2_INTERNAL_ERROR).send({ message: 'Ошибка по умолчанию' });
-    });
+    .catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -68,14 +62,9 @@ module.exports.dislikeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        return res.status(http2.constants.HTTP_STATUS_NOT_FOUND).send({ message: 'Передан несуществующий _id карточки.' });
+        throw new NotFoundError('Передан несуществующий _id карточки.');
       }
       return res.status(http2.constants.HTTP_STATUS_OK).send({ data: card });
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(http2.constants.HTTP_STATUS_BAD_REQUEST).send({ message: 'Переданы некорректные данные для постановки/снятия лайка.' });
-      }
-      return res.status(http2.constants.NGHTTP2_INTERNAL_ERROR).send({ message: 'Ошибка по умолчанию' });
-    });
+    .catch(next);
 };
